+++ app/src/main/java/com/eroom/erooja/feature/search/search_detail_page/SearchDetailPresenter.kt
@@ -1,0 +1,4 @@
+package com.eroom.erooja.feature.search.search_detail_page
+
+class SearchDetailPresenter(override val view: SearchDetailContract.View) : SearchDetailContract.Presenter {
+}
