--- app/src/main/java/com/eroom/erooja/feature/search/searchpage/SearchPresenter.kt
@@ -1,6 +1,0 @@
-package com.eroom.erooja.feature.search.searchpage
-
-import com.eroom.erooja.feature.search.searchpage.SearchContract
-
-class SearchPresenter(override val view: SearchContract.View) : SearchContract.Presenter {
-}
